requirements.txt
apache-airflow-providers-microsoft-azure
pyspark
azure-identity
azure-storage-file-datalake